requirements.txt
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.1
dj-database-url==0.4.2
Django==1.11.3
django-cors-headers==2.1.0
django-pylibmc==0.6.1
django-rest==0.0.1
django-rest-framework==0.1.0
django-rest-swagger==2.1.2
django-static-precompiler==1.6
djangorestframework==3.6.4
flake8==3.3.0
gunicorn==19.7.1
idna==2.5
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
openapi-codec==1.3.2
pprint==0.1
psycopg2==2.7.1
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
pylibmc==1.5.2
pytz==2017.2
requests==2.18.1
simplejson==3.14.0
six==1.10.0
uritemplate==3.0.0
urllib3==1.21.1
whitenoise==3.3.0
